pyproject.toml
[project]
name = "recall-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Jpsese",email = "jdumo1201@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "discord-py (>=2.6.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pynacl (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"

